removed librandom and now we're header only boys
